feat: change (mostly) routing to use camelCase i/o snake_case
